+++ chat_server/chatserver.h
@@ -14,7 +14,7 @@ public:
 #define TS_PIC_SIGN_LEN sizeof(TS_PIC_SIGN)
     ~ChatServer();
     void dispatchLine(QString line);
-    void qt_wait_ms(qint32 amount);
+    void qt_wait_ms(float amount);
     
 protected:
     void incomingConnection(int socketfd);
